Add Google Keep label to filename, export to files
The initial purpose of this fork is to add the Google Keep label to the
markdown filename and to save to markdown files instead of printing to
the terminal. It has only been tested with one label per note.
Future improvements should include support for multiple labels or
possibly saving the label names in the note header.

Considerable changes were made to the styling of the document, mostly
moving towards PEP8 guidelines. The README was also updated to match the
changes above.